dropdown component merged with searchbar component
